fixed bugs with mulitbinder

+++ src/objects/statement/cache.cc
@@ -17,7 +17,7 @@ NAN_METHOD(Statement::Cache) {
 	if (!IS_POSITIVE_INTEGER(numberValue)) {
 		return Nan::ThrowRangeError("Argument 0 must be a positive, finite integer.");
 	}
-	if (numberValue > 0x7ffffffe) {
+	if (numberValue > (double)0x7ffffffe) {
 		return Nan::ThrowRangeError("The specified cache size is too large.");
 	}
 	if (numberValue < 1) {
